[Add] Event hinzugefügt, dass ausgelöst wird, wenn ein Fehler beim Abrufen der Benachrichtigungen aufgetreten ist
+Utility.GetAnimeManga(int, Senpai) -> ProxerClass.GetAnimeManga(int,
Senpai)
+Utility-Klasse ist jetzt internal
